core: fix two more real bugs breaking the win64-native MSVC CI build
Found from the real GitHub Actions Windows job's failure log, now that the
previous three fixes let it get further than ever before:

- bitcoin-tx.vcxproj, bitcoin-util.vcxproj: unresolved GetAdjustedTimeSeconds
  (bitcoin-wallet.vcxproj: also GetMinFee). FIC's CMutableTransaction
  constructor calls GetAdjustedTimeSeconds(), and MSVC pulls in libbitcoin_
  consensus's whole transaction.obj (not just the individual functions each
  tool actually calls), making that constructor's own dependency "needed"
  even for tools that never construct a CMutableTransaction directly.
  bitcoin-tx already got the dummy stub (timedatadummy.cpp) added to its
  MSVC project in an earlier pass, matching what src/Makefile.am's
  firstislamiccoin_tx_SOURCES already does on Linux; bitcoin-util needed the
  same fix, and bitcoin-wallet needed the real implementation instead
  (src/Makefile.am's firstislamiccoin_wallet_LDADD already links
  LIBBITCOIN_NODE for this exact reason) since it also needs the real
  GetMinFee(size_t, uint32_t) from consensus/tx_verify.cpp, which only the
  node library provides.

- wallet/rpc/staking.cpp: MSVC treated a real warning as an error the first
  time this file ever compiled on Windows (only just wired into the MSVC
  project in the previous fix pass). `nTime &= ~nStakeTimestampMask` mixes
  an int64_t (nTime) with a uint32_t mask -- inverting the 32-bit mask first
  and then zero-extending it into the wider signed type would incorrectly
  clear nTime's upper 32 bits instead of just the intended low mask bits
  (harmless today since real Unix timestamps fit in 32 bits, but wrong).
  Fixed by casting the mask to int64_t before inverting, so the complement
  spans the full 64 bits nTime actually has. A different call site
  (node/miner.cpp, already fixed for UBSan during the ASan pass) ANDs the
  same mask against a genuinely uint32_t nTime field and was never affected.

Verified staking.cpp recompiles cleanly on the VPS after the fix.

+++ firstislamiccoin-core/src/wallet/rpc/staking.cpp
@@ -269,7 +269,11 @@ static RPCHelpMan checkkernel()
     CBlockIndex* pindexPrev = active_chain.Tip();
     unsigned int nBits = GetNextTargetRequired(pindexPrev, Params().GetConsensus(), true);
     int64_t nTime = GetAdjustedTimeSeconds();
-    nTime &= ~Params().GetConsensus().nStakeTimestampMask;
+    // FirstIslamicCoin: nStakeTimestampMask is uint32_t but nTime here is
+    // int64_t; casting before inverting avoids zero-extending the (32-bit)
+    // inverted mask into the wider signed type, which would incorrectly
+    // clear nTime's upper 32 bits instead of just its low mask bits.
+    nTime &= ~static_cast<int64_t>(Params().GetConsensus().nStakeTimestampMask);
 
     for (unsigned int idx = 0; idx < inputs.size(); idx++) {
         const UniValue& o = inputs[idx].get_obj();
